LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/navigate-form-userInitiated-expected.txt



FAIL <form> submission fires navigate event assert_equals: expected (string) "Submit" but got (object) null
